Refactor: Move inner classes out into separate files
